fix inf loop

+++ common/draw.c
@@ -6,7 +6,7 @@
 /*   By: flormich <flormich@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/29 10:12:54 by graja             #+#    #+#             */
-/*   Updated: 2022/01/15 10:23:31 by graja            ###   ########.fr       */
+/*   Updated: 2022/01/17 14:30:12 by graja            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -132,12 +132,11 @@ void	ft_drawTex(t_data *data, t_point p1, t_point p2, t_ray ray)
 
 	y = (float)(data->tilesize) / (p1.y - p2.y);
 	i = 0;
-	while (p1.y > p2.y)
+	while ((int)p1.y > (int)p2.y)
 	{
-			ft_draw_pixel(data, p1.x, p2.y + i, 
-					ft_getTexPixel(data, ray.offset, 
-						i * y, ray.flag));
-			i++;
-			p1.y -= 1.0;
+		ft_draw_pixel(data, p1.x, p2.y + i, 
+		ft_getTexPixel(data, ray.offset, i * y, ray.flag));
+		i++;
+		p1.y -= 1.0;
 	}
 }
